# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: El Príncipe Mestizo - Citizen journalism platform API
servers:
  - url: /api/v1
    description: Canonical API v1 path
  - url: /api
    description: Legacy (deprecated — será removido en v2)
...
paths:
...
  /articles:
    get:
      operationId: getArticles
      tags: [articles]
      summary: List published articles (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticlesPage"
...
  /articles/{slug}:
    get:
      operationId: getArticleBySlug
      tags: [articles]
      summary: Get article by slug and increment views
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{slug}/related:
    get:
      operationId: getRelatedArticles
      tags: [articles]
      summary: Get 3 related articles from the same category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
...
  /admin/articles:
    get:
      operationId: adminGetArticles
      tags: [admin]
      summary: List all articles including drafts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, published, draft]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateArticle
      tags: [admin]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/articles/{id}:
    put:
      operationId: adminUpdateArticle
      tags: [admin]
      summary: Update an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteArticle
      tags: [admin]
      summary: Delete an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/articles/{id}/publish:
    patch:
      operationId: adminPublishArticle
      tags: [admin]
      summary: Toggle publish status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/articles/{id}/feature:
    patch:
      operationId: adminFeatureArticle
      tags: [admin]
      summary: Toggle featured status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        articleCount:
          type: integer
      required:
        - id
        - name
        - slug
        - color
        - articleCount
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        coverImageAlt:
          type: ["string", "null"]
        content:
          type: string
        categoryId:
          type: integer
        secondaryCategoryId:
          type: ["integer", "null"]
        authorId:
          type: integer
        status:
          type: string
          enum: [draft, published]
        featured:
          type: boolean
        views:
          type: integer
        readingTime:
          type: integer
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        secondaryCategory:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            slug:
              type: string
            color:
              type: string
        authorName:
          type: string
      required:
        - id
        - title
        - slug
        - summary
        - content
        - categoryId
        - authorId
        - status
        - featured
        - views
        - readingTime
        - createdAt
        - updatedAt
        - category
        - authorName
    ArticlesPage:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - articles
        - total
        - page
        - limit
        - totalPages
...
    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        categoryId:
          type: integer
        coverImageUrl:
          type: ["string", "null"]
        coverImageAlt:
          type: ["string", "null"]
        featured:
          type: boolean
        status:
          type: string
          enum: [draft, published]
      required:
        - title
        - summary
        - content
        - categoryId
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        categoryId:
          type: integer
        coverImageUrl:
          type: ["string", "null"]
        coverImageAlt:
          type: ["string", "null"]
        featured:
          type: boolean
        status:
          type: string
          enum: [draft, published]
...